pyproject.toml
[project]
name = "framestormbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.32.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
